refactor: move docs-v2 to docs/agent, rename docs-new and component suffixes
