+++ apps/website/content/docs/agent/api/mock-stream-transport.mdx
