+++ apps/website/content/docs/agent/concepts/agent-architecture.mdx
