+++ apps/website/content/docs/agent/guides/streaming.mdx
