+++ apps/website/src/app/docs/[[...slug]]/page.tsx
@@ -1,8 +1,8 @@
 import { notFound } from 'next/navigation';
-import { DocsSidebarNew } from '../../../components/docs/DocsSidebarNew';
-import { MdxRendererNew } from '../../../components/docs/MdxRenderer';
+import { DocsSidebar } from '../../../components/docs/DocsSidebar';
+import { MdxRenderer } from '../../../components/docs/MdxRenderer';
 import { DocsSearch } from '../../../components/docs/DocsSearch';
-import { getDocBySlug, getAllDocSlugs } from '../../../lib/docs-new';
+import { getDocBySlug, getAllDocSlugs } from '../../../lib/docs';
 import { ApiDocRenderer, type ApiDocEntry } from '../../../components/docs/ApiDocRenderer';
 import { DocsTOC } from '../../../components/docs/DocsTOC';
 import { DocsMobileNav } from '../../../components/docs/DocsMobileNav';
@@ -12,8 +12,8 @@ import path from 'path';
 
 function loadApiDocs(): ApiDocEntry[] {
   const candidates = [
-    path.join(process.cwd(), 'apps', 'website', 'content', 'docs-v2', 'api', 'api-docs.json'),
-    path.join(process.cwd(), 'content', 'docs-v2', 'api', 'api-docs.json'),
+    path.join(process.cwd(), 'apps', 'website', 'content', 'docs/agent', 'api', 'api-docs.json'),
+    path.join(process.cwd(), 'content', 'docs/agent', 'api', 'api-docs.json'),
   ];
   for (const p of candidates) {
     if (fs.existsSync(p)) return JSON.parse(fs.readFileSync(p, 'utf8'));
@@ -46,13 +46,13 @@ export default async function DocsPage({ params }: { params: Promise<{ slug?: st
   return (
     <div className="flex min-h-screen pt-16 overflow-x-hidden" style={{ background: 'var(--gradient-bg-flow)' }}>
       <DocsSearch />
-      <DocsSidebarNew activeSection={section} activeSlug={slug} />
+      <DocsSidebar activeSection={section} activeSlug={slug} />
       <div className="flex-1 flex min-w-0" style={{ background: 'rgba(255, 255, 255, 0.85)' }}>
         <div className="flex-1 min-w-0">
           <div className="px-4 pt-4 sm:hidden">
             <DocsMobileNav activeSection={section} activeSlug={slug} />
           </div>
-          <MdxRendererNew source={doc.content} section={section} slug={slug} title={doc.title} />
+          <MdxRenderer source={doc.content} section={section} slug={slug} title={doc.title} />
           {section === 'api' && (() => {
             const entries = loadApiDocs();
             const target = API_NAME_MAP[slug];
